+++ __integration__/jest-integration/__integration__/impl/test4.feature.ts
@@ -1,3 +1,3 @@
-import { Feature } from "autometa-runner";
+import { Feature } from "@autometa/runner";
 
 Feature("../features/test4.feature");