+++ __integration__/jest-integration/CHANGELOG.md
@@ -11,5 +11,5 @@
 - Updated dependencies [554b77e]
   - @autometa/bind-decorator@0.5.0
   - @autometa/app@0.1.0
-  - autometa-runner@0.1.0
+  - @autometa/runner@0.1.0
   - @autometa/jest-transformer@0.1.0